@@ -1,3 +1,3 @@
-export const sum = (a, b) => a + b;
+export const sum = (a, b) => a + b
 
-export const sumTenToArrayItems = (arr) => arr.map((x) => sum(x, 10));
+export const sumTenToArrayItems = arr => arr.map(x => sum(x, 10))
